add back to judicial user panel SSCS-11339
